+++ doc/whats_new.md
@@ -32,9 +32,17 @@ merged into `master`! Use `git log` instead and cross-reference instead. -->
 
 - [Tushar Jamdade][]
 - [Karthikeya Kodlai][]
+- [Maira Usman][]
 
 ### Changelog
 
+- {class}`~hnn_core.Network` has new methods `Network._get_next_available_gid` and
+  `Network._shift_gid_ranges`. These are intended for future development work on
+  enabling multiple simultaneous `Network`s inside the same simulation, via enabling a
+  user to shift the GIDs of a `Network`. There is also a tutorial available in
+  `examples/howto/plot_dual_network_simulation.py`. This work was done as part of Google
+  Summer of Code 2025, organized through the International Neuroinformatics Coordinating Facility.
+  By [Maira Usman][] in {gh}`1096`.
 - Add support for printing a welcome message that asks the user to fill out the HNN
   survey, which is printed to the standard output after (and only after) the first time
   the `hnn_core` module is imported. Whether or not the message is displayed is